refactor(pdf): isolate native figure grouping
